+++ tests/conftest.py
@@ -7,6 +7,14 @@ from statham.models import ObjectSchema
 from statham.titles import title_labeller
 
 
+def pytest_configure(config):
+    # register an additional marker
+    config.addinivalue_line(
+        "markers",
+        "github_issue(id): mark test as corresponding to a github issue.",
+    )
+
+
 MAKE_LABEL = title_labeller()